More about lettuce
The five varieties of lettuce that can be grown in the garden include the following: Crisphead or Iceberg. Crisphead lettuce, more commonly known as iceberg, has a tight head of crisp leaves. Summer Crisp, French Crisp or Batavian. Butterhead, Boston or Bibb. Romaine or Cos. Looseleaf, Leaf, Cutting or Bunching.